+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Fixed iOS plugin compilation issues
+- Implemented UIImpactFeedbackGenerator for iOS haptic feedback
+
 ## 0.0.2
 
 - New release